Finished up API. Cleaned up code and organized files. Created docs.
