+++ kite-api/src/express.api.js
@@ -2,8 +2,10 @@ const express = require('express');
 const app = express();
 const cors = require('cors');
 const port = 3000;
-const docker = require('./routes/docker');
-const graphql = require('./routes/graphql');
+
+
+const docker = require('./routes/docker.router');
+const graphql = require('./routes/graphql.router');
 
 app.use('/api/docker', cors(), docker);
 app.use('/api/graphql', cors(), graphql);
