--- kite-api/src/routes/graphql.js
@@ -1,44 +1,0 @@
-const router = require('express').Router();
-const graphqlHTTP = require('express-graphql');
-const { buildSchema } = require('graphql');
-const { createContainer, stopContainer, getContainer, getAllContainers } = require('../index');
-
-const { queryType } = require('../schemas/query.schema');
-const { mutationType } = require('../schemas/mutation.schema');
-const { Container, containerType } = require('../schemas/container.schema');
-
-const schema = buildSchema(`
-    ${containerType}
-    ${mutationType}
-    ${queryType}
-`);
-
-const root = {
-    createContainer: async ({userid}) => {
-        try {
-            await createContainer(userid);
-            return `Successfully created container for student ${userid}`;
-        } catch (err) {
-            return `Error creating container for student - ${err}`;
-        }
-    },
-    getContainer: async ({userid}) => {
-        try {
-            let container = await getContainer(`${userid}php`);
-            return container;
-        } catch (err) {
-            return `Error retrieving your container - ${err}`;
-        }
-    },
-    hello: () => {
-        return 'Hello World!';
-    }
-};
-
-router.use('/', graphqlHTTP({
-    schema: schema,
-    rootValue: root,
-    graphiql: true,
-}));
-
-module.exports = router;
