+++ kite-api/src/routes/graphql.router.js
@@ -1,0 +1,26 @@
+const router = require('express').Router();
+const graphqlHTTP = require('express-graphql');
+const { buildSchema } = require('graphql');
+
+const { queryType } = require('../schemas/query.schema');
+const { mutationType } = require('../schemas/mutation.schema');
+const { containerType } = require('../schemas/container.schema');
+
+const schema = buildSchema(`
+    ${containerType}
+    ${mutationType}
+    ${queryType}
+`);
+
+const { ContainerResolvers } = require('../resolvers/container.resolver');
+
+const root = {};
+Object.assign(root, ContainerResolvers);
+
+router.use('/', graphqlHTTP({
+    schema: schema,
+    rootValue: root,
+    graphiql: true,
+}));
+
+module.exports = router;
