+++ kite-api/src/schemas/mutation.schema.js
@@ -1,6 +1,7 @@
 const mutationType = `
     type Mutation {
         createContainer(userid: String!): String
+        deleteContainer(userid: String!): String
     }
 `;
