+++ kite-api/src/schemas/query.schema.js
@@ -1,7 +1,8 @@
 const queryType = `
     type Query {
+        getContainer(userid: String!): Container!
+        getAllContainers: [Container!]!
         hello: String
-        getContainer(userid: String!): Container
     }
 `;
